yelp_backend_v2/user_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
python-multipart
pillow
langchain
langchain-groq
langchain-community
tavily-python
python-dotenv
pydantic[email]
cryptography
pymongo
motor
aiokafka
